feat: auto auditing
