+++ middleware/wrapper/index.js
@@ -1,7 +1,5 @@
 
-const extractAuditData = require('./extractAuditData');
 module.exports = ({port, options}) => {
-    const {audit} = options;
     return async(ctx, next) => {
         // request
         ctx.ut = {};
@@ -14,6 +12,5 @@ module.exports = ({port, options}) => {
             if (port.config.debug) ctx.body.error.debug = {stack: e.stack.split('\n')};
             ctx.app.emit('error', e, ctx);
         }
-        if (audit && ctx.ut.method) ctx.app.emit('audit', extractAuditData(port, ctx));
     };
 };
